fix: move apply patch tool to tools/experimentals
