+++ .github/CODEOWNERS
@@ -7,3 +7,6 @@
 /test/a2a/ @Lancetnik
 /test/remote/ @Lancetnik
 /website/docs/user-guide/a2a/ @Lancetnik
+
+# GPT 5.1 Apply Patch Tool
+/autogen/tools/experimental/apply_patch/ @priyansh4320
