+++ MAINTAINERS.md
@@ -25,7 +25,7 @@
 | George Sideris    | [giorgossideris](https://github.com/giorgossideris)        | -                        | reasoning, RAG                          |
 | Beibin Li         | [BeibinLi](https://github.com/BeibinLi)                    | GenAI, Meta              | multimodal, reasoning, optiguide        |
 | Nikita Pastukhov  | [Lancetnik](https://github.com/lancetnik)                  | -                        | A2A                                     |
-
+| Priyanshu Deshmukh| [priyansh4320](https://github.com/priyansh4320)            | -                        | GPT-5 support,apply patch tool          |
 
 **Pending Maintainers list (Marked with \*, Waiting for explicit approval from the maintainers)**
 | Name            | GitHub Handle                                              | Organization           | Features                                |
