+++ autogen/tools/experimental/apply_patch/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2023 - 2025, AG2ai, Inc., AG2ai open-source projects maintainers and core contributors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from .apply_patch_tool import ApplyPatchTool, PatchEditor, WorkspaceEditor
+
+__all__ = ["ApplyPatchTool", "PatchEditor", "WorkspaceEditor"]
